% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uni_gpa.R
\name{gpa_calculator}
\alias{gpa_calculator}
\title{Interactive GPA calculator with memory and updates}
\usage{
gpa_calculator()
}
\value{
Displays updated GPA and stores the data for future use.
}
\description{
Loads saved GPA data if available. You can enter new grades + ECTS to update your GPA.
If no new input is provided, it simply shows your current GPA.
}
\examples{
\dontrun{
gpa_calculator()
}
}
