% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motivate.R
\name{motivate_student}
\alias{motivate_student}
\title{Motivate a student with a random message}
\usage{
motivate_student()
}
\value{
A motivational message (character)
}
\description{
Motivate a student with a random message
}
\examples{
motivate_student()
}
