% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reset_gpa_data.R
\name{reset_gpa_data}
\alias{reset_gpa_data}
\title{Reset GPA data file}
\usage{
reset_gpa_data(file = "gpa_data.csv", gpa_file = "gpa_info.csv")
}
\arguments{
\item{file}{The name of the GPA data file (default: "gpa_data.csv")}

\item{gpa_file}{The name of the GPA value file (default: "gpa_info.csv")}
}
\value{
Message indicating success or cancellation.
}
\description{
Deletes the current GPA data files ("gpa_data.csv" and "gpa_info.csv") after confirmation.
}
\examples{
\dontrun{
reset_gpa_data()
}
}
